requirements.txt
pyside6==6.7.1
openai>=1.3.0
pillow
numpy
pynput
